requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP Client for Azure DevOps API
httpx==0.26.0
python-multipart==0.0.6

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Background Tasks
celery==5.3.6
redis==5.0.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0.post1

# Monitoring & Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
